pyproject.toml
[tool.poetry]
name = "lesson27_solution"
version = "0.1.0"
description = ""
authors = ["Albina <aalmukhametova@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.1"
Pillow = "^9.0.0"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.2"
django-pytest = "^0.2.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
pytest = "^7.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
